Moving ChatMessages to separate component
